requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymongo==4.8.0
motor==3.5.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
python-dotenv==1.0.1
apscheduler==3.10.4
httpx==0.27.2
